# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TICKET API",
        "description": "Documentation de mon API Laravel",
        "contact": {
            "name": "Support API",
            "email": "shadracmeshacdimon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-intents": {
            "get": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Récupérer toutes les intentions de commande",
                "operationId": "3b95437fa58cf89d18c384277358a46d",
                "responses": {
                    "200": {
                        "description": "Liste des intentions de commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderIntent"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Créer une nouvelle intention de commande",
                "operationId": "6237f7cc81ff48ca3b38154d748e2f69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_intent_price",
                                    "order_intent_type",
                                    "user_email",
                                    "user_phone",
                                    "expiration_date"
                                ],
                                "properties": {
                                    "order_intent_price": {
                                        "type": "integer"
                                    },
                                    "order_intent_type": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "user_email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 100
                                    },
                                    "user_phone": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Intention de commande créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderIntent"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order-intents/{id}": {
            "get": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Récupérer une intention de commande par ID",
                "operationId": "8a88699e9e0f66499d58ce239b516504",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'intention de commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Intention de commande trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderIntent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Intention de commande non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order-intents/clean-expired": {
            "post": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Nettoyer les intentions de commande expirées",
                "operationId": "ced06ef631c355c944339b77209667fb",
                "responses": {
                    "200": {
                        "description": "Intentions de commande expirées nettoyées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "deleted_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/validate-order-intent/{id}": {
            "post": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Valider une intention de commande",
                "operationId": "e3644d47645441c08d9b4c106155ceed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'intention de commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande validée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "order_number": {
                                            "type": "string"
                                        },
                                        "ticket_download_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de la validation de la commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderIntent": {
                "properties": {
                    "order_intent_id": {
                        "type": "integer"
                    },
                    "order_intent_price": {
                        "type": "integer"
                    },
                    "order_intent_type": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "expiration_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Use the token provided after logging in.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}